Fixed logic error causing AM/PM to not read correctly
I use military time for most logic and only really convert it to the am/pm format for displaying. I forgot to fix the underflow issue caused by changing timezones for the military time outside of the convertTimeFormat function. I moved that code to fix the underflow outside of the function.