Changed the pins for the lcd to allow for an override button
Needed to move the pins over one because I needed pin 2 for an interrupt. I set pin 2 to interrupt when pin 2 see's a change in state from ground to disconnected. On an interrupt event the interrupt calls overrideRelay. It seems to work as intended. I need to now find a button that holds its position when pressed.