Made a change to the sunrise sunset checker logic
I had a problem where the logic was checking for both the hour and minute to be less than or greater than. the problem is every hour the minute resets causing the if statement to fail when it shouldn't. 7:58pm is valid and turns on but 8:00pm is not valid and turns off when it should be valid. This is now fixed