Added DST accountability and commenting blocks
added code to convertTimeZone to request a server if DST is currently happening and account for it in converting the time zone. I also added the header block of commenting describing the project and giving credit to the resources used. Each function has a comment block on the top describing the functions purpose, modification date, return types, etc.